@@ -1,119 +1,119 @@
 {
-  "name": "angular",
-  "version": "2.0.0-alpha.21",
+  "name": "angular-srcs",
+  "version": "2.0.0-rc.4",
+  "private": true,
+  "branchPattern": "2.0.*",
   "description": "Angular 2 - a web framework for modern web apps",
   "homepage": "https://github.com/angular/angular",
   "bugs": "https://github.com/angular/angular/issues",
-  "contributors": [
-    "Alex Eagle <alexeagle@google.com>",
-    "Chirayu Krishnappa <chirayu@google.com>",
-    "Jeff Cross <crossj@google.com>",
-    "Misko Hevery <misko@google.com>",
-    "Rado Kirov <radokirov@google.com>",
-    "Tobias Bosch <tbosch@google.com>",
-    "Victor Savkin <vsavkin@google.com>",
-    "Yegor Jbanov <yjbanov@google.com>"
-  ],
-  "license": "Apache-2.0",
+  "license": "MIT",
+  "engines": {
+    "node": ">= 5.4.1 < 6"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/angular/angular.git"
   },
   "scripts": {
-    "postinstall": "webdriver-manager update && bower install && gulp pubget.dart && tsd reinstall --config modules/angular2/tsd.json && tsd reinstall --config tools/tsd.json",
-    "test": "gulp test.all.js && gulp test.all.dart"
+    "postinstall": "node tools/npm/copy-npm-shrinkwrap && webdriver-manager update"
   },
   "dependencies": {
-    "es6-module-loader": "^0.9.2",
-    "googleapis": "1.0.x",
-    "gulp-insert": "^0.4.0",
-    "gulp-modify": "0.0.4",
-    "gulp-replace": "^0.5.3",
-    "node-uuid": "1.4.x",
-    "reflect-metadata": "^0.1.0",
-    "rx": "2.5.1",
-    "selenium-webdriver": "2.45.1",
-    "systemjs": "^0.9.1",
-    "traceur": "0.0.87",
-    "which": "~1",
-    "zone.js": "0.4.3"
+    "es6-shim": "^0.35.0",
+    "reflect-metadata": "^0.1.3",
+    "rxjs": "5.0.0-beta.6",
+    "zone.js": "^0.6.12"
   },
   "devDependencies": {
-    "angular": "1.3.5",
+    "@types/angularjs": "^1.5.13-alpha",
+    "@types/fs-extra": "0.0.22-alpha",
+    "@types/hammerjs": "^2.0.22-alpha",
+    "@types/jasmine": "^2.2.22-alpha",
+    "@types/node": "^4.0.22-alpha",
+    "@types/protractor": "^1.5.11-alpha",
+    "@types/selenium-webdriver": "^2.44.20-alpha",
+    "angular": "^1.5.0",
+    "angular-animate": "^1.5.0",
+    "angular-mocks": "^1.5.0",
+    "base64-js": "^0.0.8",
     "bower": "^1.3.12",
-    "broccoli": "^0.15.3",
-    "broccoli-filter": "^0.1.12",
-    "broccoli-flatten": "^0.1.1",
-    "broccoli-funnel": "igorminar/broccoli-funnel#perf-files",
-    "broccoli-lodash": "^0.1.1",
-    "broccoli-merge-trees": "^0.2.1",
-    "broccoli-replace": "alexeagle/broccoli-replace#angular_patch",
-    "broccoli-slow-trees": "^1.1.0",
+    "broccoli": "^0.16.9",
+    "broccoli-file-creator": "^1.1.0",
+    "broccoli-funnel": "^1.0.1",
+    "broccoli-slow-trees": "1.x.x",
     "broccoli-stew": "^0.2.1",
     "broccoli-writer": "^0.1.1",
+    "browserify": "^10.2.4",
+    "browserstacktunnel-wrapper": "^1.4.2",
     "canonical-path": "0.0.2",
-    "conventional-changelog": "^0.0.17",
-    "css": "mlaval/css#issue65",
-    "del": "~1",
-    "dgeni": "^0.4.1",
-    "dgeni-packages": "^0.10.11",
-    "event-stream": "^3.1.5",
-    "fs-extra": "^0.18.0",
+    "chokidar": "^1.1.0",
+    "clang-format": "^1.0.32",
+    "conventional-changelog": "^1.1.0",
+    "cors": "^2.7.1",
+    "firefox-profile": "^0.3.4",
+    "fs-extra": "^0.26.3",
     "glob": "^4.0.6",
     "gulp": "^3.8.8",
-    "gulp-autoprefixer": "^2.1.0",
-    "gulp-changed": "^1.0.0",
-    "gulp-clang-format": "^1.0.3",
+    "gulp-autoprefixer": "^3.1.0",
+    "gulp-clang-format": "^1.0.23",
     "gulp-concat": "^2.5.2",
-    "gulp-connect": "~1.0.5",
-    "gulp-load-plugins": "^0.7.1",
+    "gulp-connect": "^2.3.1",
+    "gulp-conventional-changelog": "^1.1.0",
+    "gulp-gzip": "^1.2.0",
+    "gulp-insert": "^0.5.0",
+    "gulp-load-plugins": "^1.0.0",
     "gulp-rename": "^1.2.0",
-    "gulp-sass": "^1.3.3",
-    "gulp-shell": "^0.2.10",
-    "gulp-sourcemaps": "1.3.*",
-    "gulp-template": "^3.0.0",
-    "gulp-traceur": "0.17.*",
-    "gulp-typescript": "^2.6.0",
-    "gulp-webserver": "^0.8.7",
-    "html2jade": "^0.8.3",
-    "indent-string": "^1.2.1",
-    "js-beautify": "^1.5.5",
+    "gulp-replace": "^0.5.3",
+    "gulp-sass": "^2.0.4",
+    "gulp-shell": "^0.5.1",
+    "gulp-sourcemaps": "^1.6.0",
+    "gulp-tslint": "^5.0.0",
+    "gulp-typescript": "^2.10.0",
+    "gulp-uglify": "^1.2.0",
+    "hash-files": "^1.0.0",
+    "jasmine": "^2.4.1",
+    "jpm": "1.0.0",
     "js-yaml": "^3.2.7",
-    "karma": "^0.12.23",
-    "karma-chrome-launcher": "^0.1.4",
-    "karma-cli": "^0.0.4",
-    "karma-dart": "^0.2.8",
-    "karma-jasmine": "^0.2.2",
-    "lodash": "^2.4.1",
-    "madge": "^0.5.0",
-    "marked": "^0.3.3",
+    "karma": "^0.13.20",
+    "karma-browserstack-launcher": "^0.1.9",
+    "karma-chrome-launcher": "^0.2.0",
+    "karma-jasmine": "^0.3.6",
+    "karma-sauce-launcher": "^0.3.0",
+    "karma-sourcemap-loader": "^0.3.6",
+    "lodash": "^3.10.1",
+    "madge": "0.5.0",
     "merge": "^1.2.0",
     "merge2": "^0.3.5",
-    "minijasminenode2": "^1.0.0",
-    "minimatch": "^2.0.1",
-    "minimist": "1.1.x",
-    "mock-fs": "^2.5.0",
-    "node-html-encoder": "0.0.2",
+    "minimatch": "^3.0.0",
+    "minimist": "^1.2.0",
+    "mock-fs": "^3.6.0",
+    "node-uuid": "1.4.x",
+    "on-headers": "^1.0.0",
     "parse5": "1.3.2",
-    "protractor": "2.0.0",
+    "protractor": "^3.0.0",
+    "proxy-middleware": "^0.15.0",
     "q": "^1.0.1",
-    "react": "^0.13.2",
-    "run-sequence": "^0.3.6",
+    "react": "^0.14.0",
+    "rewire": "^2.3.3",
+    "rollup": "^0.26.3",
+    "run-sequence": "^1.1.0",
+    "selenium-webdriver": "^2.47.0",
+    "semver": "^5.1.0",
     "sorted-object": "^1.0.0",
     "source-map": "^0.3.0",
-    "sprintf-js": "1.0.*",
-    "string": "^3.1.1",
+    "source-map-loader": "^0.1.5",
+    "strip-ansi": "^2.0.1",
     "symlink-or-copy": "^1.0.1",
+    "systemjs": "0.18.10",
     "systemjs-builder": "^0.10.3",
-    "temp": "^0.8.1",
-    "ternary-stream": "^1.2.3",
-    "through2": "^0.6.1",
-    "ts2dart": "^0.4.8",
-    "tsd": "^0.5.7",
-    "typescript": "alexeagle/TypeScript#error_is_class",
-    "vinyl": "^0.4.6",
-    "walk-sync": "^0.1.3",
-    "xtend": "^4.0.0",
-    "yargs": "2.3.*"
+    "through2": "^0.6.5",
+    "ts-api-guardian": "0.1.4",
+    "ts2dart": "^0.9.10",
+    "tsickle": "^0.1.4",
+    "tslint": "^3.10.0-dev.2",
+    "typescript": "^1.9.0-dev.20160409",
+    "universal-analytics": "^0.3.9",
+    "webpack": "^1.12.6",
+    "which": "~1",
+    "yargs": "^3.31.0"
   }
 }
